requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
# psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
openpyxl==3.1.2
pandas==2.1.1
reportlab==4.0.4
whitenoise==6.5.0
gunicorn==21.2.0
